Clone function created for all Ship classes
This is needed to help the copy from one Field to another, as it have a Ship* list (polymorphism), I dont want to share Ships resources between the maps.

Should be noted that there must have a Clone function for each derived from Ship class, as Ship can't be cloned.